Convert to int

+++ neurons/deployment_layer/model_fa0d509d52abe2d1e809124f8aba46258a02f7253582f7b7f5a22e1e0bca0dfb/input.py
@@ -84,7 +84,7 @@ class CircuitInput(BaseInput):
             "minimum_response_time": [minimum_response_time for _ in range(BATCH_SIZE)],
             "maximum_response_time": [maximum_response_time for _ in range(BATCH_SIZE)],
             "response_time": [response_time for _ in range(BATCH_SIZE)],
-            "competition": [random.random() * SCALING for _ in range(BATCH_SIZE)],
+            "competition": [int(random.random() * SCALING) for _ in range(BATCH_SIZE)],
             "scaling": SCALING,
             "RATE_OF_DECAY": int(RATE_OF_DECAY * SCALING),
             "RATE_OF_RECOVERY": int(RATE_OF_RECOVERY * SCALING),
